audio player structured into a component and reuse
